requirements.txt
google-api-core==2.8.1
google-auth==2.6.6
google-cloud-core==2.3.0
google-cloud-datastore==2.6.1
google-cloud-logging==3.1.1
google-cloud-storage==2.11.0
googleapis-common-protos==1.56.2
requests==2.27.1
urllib3==1.26.9
pytz==2023.3
openai==1.13.3